rake move_game[g_1860]
